Conflicts:
	CMake/ctkMacroGenerateMocs.cmake
	CMake/ctkMacroSetupQt.cmake
	Libs/Core/Testing/Cpp/CMakeLists.txt
	SuperBuild.cmake

+++ Libs/Core/ctkBackTrace.cpp
@@ -73,7 +73,7 @@ struct ctkBackTracePrivate
 {
   std::vector<void *> Frames;
 
-  int trace(void** addresses, int size) const;
+  int trace(void** addresses, size_t size) const;
   std::string getSymbol(void* address) const;
 };
 
@@ -89,7 +89,7 @@ ctkBackTrace::ctkBackTrace(size_t framesNumber)
   if(framesNumber == 0)
     return;
   d->Frames.resize(framesNumber, 0);
-  int size = d->trace(&d->Frames.front(), framesNumber);
+  size_t size = d->trace(&d->Frames.front(), framesNumber);
   d->Frames.resize(size);
 }
 
@@ -147,7 +147,7 @@ QList<QString> ctkBackTrace::stackTrace() const
 #if defined(CTK_HAVE_EXECINFO)
 
 // --------------------------------------------------------------------------
-int ctkBackTracePrivate::trace(void** array, int n) const
+int ctkBackTracePrivate::trace(void** array, size_t n) const
 {
   return :: backtrace(array,n);
 }
@@ -155,7 +155,7 @@ int ctkBackTracePrivate::trace(void** array, int n) const
 #elif defined(Q_CC_MSVC)
 
 // --------------------------------------------------------------------------
-int ctkBackTracePrivate::trace(void** array, int n) const
+int ctkBackTracePrivate::trace(void** array, size_t n) const
 {
   if(n>=63)
     n=62;
@@ -165,7 +165,7 @@ int ctkBackTracePrivate::trace(void** array, int n) const
 #else
 
 // --------------------------------------------------------------------------
-int ctkBackTracePrivate::trace(void** /*array*/, int /*n*/) const
+int ctkBackTracePrivate::trace(void** /*array*/, size_t /*n*/) const
 {
   return 0;
 }
